Add hamburger-menu layout and animation on click
Implemented hamburger to cross layout animation of hamburger buns
this is achieved by listening to the status of the checkbox input.
on checked the top-bun and bottom-bun margins and rotation
will be changed. the transition of the changes have been delayed
to start with the margin and end with rotation on input:checked
and do the same in reversed order if unchecked again.