Add position and z-index to navigation-links elements Also keyframes
Added position of relative to navigation-link-group, navigation-dropdown and
navigation-link-group ul to create a stacking order. added overflow hidden
to the two top layers to hide any overflow from the underneath ul element

Added a keyframes animation that starts at a 6rem higher y axis and ends
on the original element position.

this appearOnDropDown animation has been added to any ul that is
a child of a navigation-link-group element that also has the
dropdown-expanded class. this creates an 'appear on dropdown' animation
for the ul element