Add initial layout and styles for apple-tv section elements.
mainly added heights and widths and initialized flex container layouts.
added gradient-overlay-dark and .image-carousel-container-active which
together can trigger an 'appearing from the bottom' animation
on carousel-item-info-bottom elements. effectively getting triggered
and starting an 'appearing' transition if the a carousel-item is the main
item currently, and will trigger a 'disappearing' reverse transition if
the current carousel-item is not the current active item anymore