Implement carousel scrolling through adjusting --scroll-distance root variable
Add the function scrollImageCarousel that takes in one argument for the direction
either 'left' or 'right'.
It retrieves the current '--scroll-distance' value and adds or substracts
the width of a carousel-item-container.

the '--scroll-distance' custom css variable influences the transform
translate-x value of the carousel-item-container elements, which results
in a scrolling transition.