Implement logic to handle scroll to next or previous image and infinite carousel
- Add logic to scroll to first image-carousel position
- Add logic to scroll to last image-carousel position

both if statements add the 'image-carousel-container-active' class
to their desired container data-index values and then remove the previous
active class from the last container.

scrolling to 'first' position resets the --scroll-distance to 0
scrolling to 'last' decreases the scroll-distance value by the container width times 10

- Add wrapper function handleScrollToLast and handleScrollToFirst
they are used as callbacks for the transitionend eventlisteners
will trigger the scrollImageCarousel with either 'first' or 'last' as argument

- Add handleScrollToLeft and handleScrollToRight
both functions are set to add and remove the 'image-carousel-container-active' class
to the new active container and from the previously active container.
If the new current active container has a data-index value ending in 'c',
which marks copy containers, used for the infinity scroll setup, the functions
will add a transitionend listener to the first carousel-item-container to wait
for the translate-x transition to end, to call the handleScrollToLast or
handleScrollToFirst functions, which are designed to reset the positioning of the containers
unnoticed by disabling the animations inside apple-tv and then turning them back on after

- Add imageCarouselClickHandler to delegate click events inside apple-tv
current setup adds logic for the previous and next buttons.

- Add debounceCarouselClickHandler
removes the imageCarouselClickHandler eventlistener and sets a timeout for 1001ms
to then reattach the same eventlistener back. This helps mitigating bugs if a user tries to
violently trigger the scroll animation repeatedly