Add function to adjust scrollDistance after resize. Add Debouncer
- Implement debouncer debounce(func, wait)
takes in a callback and a wait time.
delays the execution of the given callback for the duration of the wait time,
starting from the last time the debouncer has been invoked. if its invoked while
the wait time isnt over, the wait time will start again until passed without further
calls on the debouncer

- Implement adjustImageCarouselOnResize
calculates the correct value for --scroll-distance, a custom css variable that
is used to set the translate-x property of carousel items, to create a smooth
scrolling animation. At certain media breakpoints the size of the carousel items changes
and thus the --scroll-distance needs to be recalculated to avoid misaligning of the carousel
items.

- Add resize eventlistener to window that triggers a debounced adjustImageCarouselOnResize