Implement flyout height calculation based on content and nav-item flyout handling
- addNavItemFlyoutListeners
 Using the nodelist for all nav-item elements, each nav-item get an mouseenter
eventlistener attached to it, which triggers the toggleFlyoutItemVisibility on the
matching index flyout-item element from the flyout-item nodelist
after every matching index nav-item and flyout-item are paired,
a mouseleave eventlistener that calls flyoutClose, which closes any flyout elements,
is attached to the flyoutContainer and the appleIcon

- flyoutClose
Function that hides all flyout-container children and removes any added flyout related classes
and states

- flyoutOpen
Adds flyout-open class to navbar and flyoutContainer. checks the hamburgermenu checkbox

- toggleFlyoutItemVisibility
tailored for flyout-item flyouts, multiple functions for opening the flyoutContainer and
hiding other flyout elements are called. Adjusts the flyoutContainers height by manipulating
the --flyout-height css variable

- saveNavItemFlyoutHeightIndicator
Iterates over every element of the flyoutItems nodelist.
it counts the list items of each unordered list inside the flyout element
and counts the first unordered lists list items 1 for 1 and all subsequently ul li for half the value
this is in line with the first ul having double the font size of all other ul inside the same flyout.
after comparing and getting the highest counted value, the value then gets passed into a height calculation
function and its returned value  saved as data-flyout-height
on the flyout element.

- calculateFlyoutHeight
takes an argument of listItemNumber
retrieves all flyout-container related css variables needed to calculate the optimal height
and then calculates the optimal height for the flyout element based on its highest listitemNumber

- adjustFlyoutContainerHeight
sets the --flyout-height css variable to the value of the currently active flyout elements
 data-flyout-height value in pixel