Add clicklisteners to flyout-mobile-item's and mobileBackNavigation
- mobileItemToFlyoutItem
gets passed a flyoutItem
only runs if the viewports width is less than 833px
classes will be applied or removed to make the passed in flyoutItem
fade in from the right.

- mobileItemBackNavigationHandler
resets the flyoutItems and itself.
adds the navigatedBack class to flyoutMobile and turns it visible
this results in the flyoutMobile fading in from the left

- addFlyoutMobileItemsClickListeners
Loops over all flyout-mobile-item elements and adds a click listener
to each with their index paired flyoutitem and finally adds a click
listener to mobileBackNavigation